tested

+++ backend/src/main/java/com/codestates/pre31/qna/question/DTO/PostQuestionDTO.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Data
 public class PostQuestionDTO {
 
-    private Integer user_id;
+    private long user_id;
 
     private String title;
